Document heap/tracker as single-thread and split pool internals.
Heap and heap tracker are single-thread only: create, use, and
release on the same thread (normally one compiled_value evaluate).
They now always use the single-thread inf. afw_pool_internal.h is
shared decls only; heap structs moved to afw_pool_heap_internal.h;
the general-pool instance struct lives in afw_pool.c.
