+++ src/afw/pool/afw_pool_internal.h
@@ -18,171 +18,20 @@
 
 /**
  * @file afw_pool_internal.h
- * @brief Adaptive Framework memory pool internal header.
+ * @brief Shared pool internals for libafw (create base pool, debug).
+ *
+ * Heap/tracker structs live in afw_pool_heap_internal.h (afw_pool_heap.c
+ * only). The general-pool instance struct lives in afw_pool.c.
  */
 
 AFW_BEGIN_DECLARES
-
-typedef struct afw_pool_internal_inf_implementation_specific_s {
-    afw_boolean_t is_multithreaded;
-    afw_boolean_t is_subpool;
-} afw_pool_internal_inf_implementation_specific_t;
-
-typedef struct afw_pool_internal_memory_prefix_s
-afw_pool_internal_memory_prefix_t;
-
-/**
- * @brief Memory prefix
- * 
- * This is the prefix before each address returned by afw_pool_calloc()
- * and afw_pool_malloc() for implementations that do not keep a chain of
- * allocated memory.
- */
-struct afw_pool_internal_memory_prefix_s {
-    afw_size_t size;
-    const afw_pool_t *p;
-    /* Allocated/free memory starts here. */
-};
-
-#define AFW_POOL_INTERNAL_MEMORY_PREFIX(address) \
-    (afw_pool_internal_memory_prefix_t *) \
-    (((char *)address) - sizeof(afw_pool_internal_memory_prefix_t))
-
-typedef struct afw_pool_internal_memory_prefix_with_links_s
-afw_pool_internal_memory_prefix_with_links_t;
-
-/*
- * @brief Memory prefix with links
- *
- * This is the prefix before each address returned by afw_pool_calloc()
- * and afw_pool_malloc() for implementations that keep up with a chain of
- * allocated memory like the subpool* implementations.
- */
-struct afw_pool_internal_memory_prefix_with_links_s {
-    afw_pool_internal_memory_prefix_with_links_t *prev;
-    afw_pool_internal_memory_prefix_with_links_t *next;
-    /* Common prefix must always be at end. */
-    afw_pool_internal_memory_prefix_t common;
-    /* Allocated/free memory starts here. */
-};
-
-#define AFW_POOL_INTERNAL_MEMORY_PREFIX_WITH_LINKS(address) \
-    (afw_pool_internal_memory_prefix_with_links_t *) \
-    (((char *)address) - sizeof(afw_pool_internal_memory_prefix_with_links_t))
-
-
-typedef struct afw_pool_internal_free_memory_s
-afw_pool_internal_free_memory_t;
-
-struct afw_pool_internal_free_memory_s {
-    afw_pool_internal_free_memory_t *next;
-    afw_size_t size;
-    /* Free memory starts here. */
-};
-
-
-typedef struct afw_pool_internal_free_memory_head_s
-afw_pool_internal_free_memory_head_t;
-/** @brief The head of each free memory free_block. */
-struct afw_pool_internal_free_memory_head_s {
-    afw_pool_internal_free_memory_t *first; /* This will go away. */
-};
-
-typedef struct afw_pool_internal_self_s
-afw_pool_internal_self_t;
-
-struct afw_pool_internal_self_s {
-
-    afw_pool_t pub;
-
-    /** @brief Unique number for pool. */
-    afw_integer_t pool_number;
-
-    /** @brief Associated apr pool. This might be the same a parent's. */
-    apr_pool_t *apr_p;
-
-    /**
-     * @brief Public apr pool or NULL if not exposed yet.
-     * 
-     * This is set the first time when afw_pool_get_apr_pool() is called.
-     * If this pool allocates from parent, a new pool is created if needed. If
-     * not, public_apr_p is set to apr_p.
-     */
-    apr_pool_t *public_apr_p;
-
-    /** @brief Optional pool name. */
-    const afw_utf8_t *name;
-
-    /** @brief Parent heap when this is a tracker. */
-    afw_pool_internal_self_t *parent;
-
-    /**
-     * @brief AFW parent when this is a heap (usually an APR pool).
-     */
-    const afw_pool_t *external_parent;
-
-    /** @brief First subpool of this pool. */
-    afw_pool_internal_self_t *first_child;
-
-    /** @brief Next sibling of this pool. */
-    afw_pool_internal_self_t *next_sibling;
-
-    /**
-     * @brief Thread associated with a thread specific pool.
-     *
-     * If this is not NULL, this pool is thread specific and can only be
-     * accessed by this thread.
-     */
-    const afw_thread_t *thread;
-
-    /** @brief First cleanup function. */
-    afw_pool_cleanup_t *first_cleanup;
-
-    /**
-     * @brief Pools reference count.
-     *
-     * This starts at 1 on create and is incremented and decremented
-     * by afw_pool_get_reference() and afw_pool_release().
-     */
-    afw_integer_t reference_count;
-
-    /** @brief Bytes allocated via afw_pool_malloc()/afw_pool_calloc(). */
-    afw_size_t bytes_allocated;
-
-    /**
-     * @brief First allocated memory.
-     * 
-     * This will be NULL if implementation does not keep up with allocated
-     * memory.
-     */
-    afw_pool_internal_memory_prefix_with_links_t *first_allocated_memory;
-
-    /**
-     * @brief Free memory head.
-     * 
-     * For subpool implementations, this is the same as the parent's.
-     */
-    afw_pool_internal_free_memory_head_t *free_memory_head;
-};
-
-
-typedef struct afw_pool_internal_self_with_free_memory_head_s
-afw_pool_internal_self_with_free_memory_head_t;
-struct afw_pool_internal_self_with_free_memory_head_s {
-
-    afw_pool_internal_self_t common;
-
-    /* Don't assess this directly. Use free_memory_head pointer instead. */
-    afw_pool_internal_free_memory_head_t memory_for_free_memory_head;
-};
-
 
 /**
  * @internal
  * @brief Allocate base pool.
  * @return Pointer to base pool or NULL if there is an error.
  *
- * This should only be used in afw_environment_create().
+ * Multithreaded. Only used in afw_environment_create().
  */
 AFW_DECLARE(const afw_pool_t *)
 afw_pool_internal_create_base_pool();
@@ -197,6 +46,6 @@ afw_pool_print_debug_info(
 
 AFW_END_DECLARES
 
-/** @} */  // end of @addtogroup @addtogroup
+/** @} */
 
 #endif /* __AFW_POOL_INTERNAL_H__ */
